git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/key-based-refactor@491355 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-store/src/test/java/org/apache/maven/continuum/store/utils/StoreTestUtils.java
@@ -328,7 +328,7 @@ public class StoreTestUtils
         return dependency;
     }
 
-    public static ProjectDeveloper createTestDeveloper( long continuumId, String email, String name, String scmId )
+    public static ProjectDeveloper createTestDeveloper( int continuumId, String email, String name, String scmId )
     {
         ProjectDeveloper developer = new ProjectDeveloper();
         developer.setContinuumId( continuumId );
